Fixed 2011b incompatibility; changed HL-20 model location
Simulink generation script was causing an error, "Attempt to modify 'simulink'
which is a locked (read-only) library." Fixed by explicitly using name of
model under test instead of relying on gcs to reply with name of most recently
opened system.

Changed location of HL-20 DAVE-ML model to reflect new structure on my
computer; this change is in test/app-tests/Makefile, not the Makefile.dist
version.